fix: Correct count argument to be type uint64 instead of int

+++ internal/data/model/pets.go
@@ -99,7 +99,7 @@ func (p *PetModel) GetPetCount(id *string) (int, error) {
 	return count, err
 }
 
-func (p *PetModel) UpdatePetCount(petID string, userID string, count int) error {
+func (p *PetModel) UpdatePetCount(petID string, userID string, count uint64) error {
 	query := `INSERT INTO UserPetsClickCount (pet_id, user_id, click_count) VALUES ($1, $2, $3) ON CONFLICT (pet_id, user_id) DO UPDATE SET click_count = UserPetsClickCount.click_count + $3;`
 	_, err := p.DB.Exec(query, petID, userID, count)
 	return err
